working on site redesign using navbar and routes
