+++ package.json
@@ -11,6 +11,7 @@
     "react-bootstrap": "^2.5.0",
     "react-draggable": "^4.4.5",
     "react-native-web": "^0.18.9",
+    "react-router-dom": "^6.14.2",
     "react-scripts": "^5.0.1",
     "serve": "^14.0.1",
     "web-vitals": "^3.0.1"
